Fix uploading from url
The response for checking the status of an upload via url seems to have
changed from what this library expected. The `status` parameter was
returning `progress` which stopped the while loop but did not return a
`file_id` is it was not yet available. This commit switches the logic to
explicitly wait for `success`.

In addition, I've added a `MAX_RETRY_ATTEMPTS` constant and extracted the
retry frequency into a `RETRY_AFTER` constant.

Finally, I use `Hash#fetch` to get the `file_id` from the response to
prevent a silent error in the case that the response does not contain an
id if the MAX_RETRY_ATTEMPTS is exceeded or something else unexpected
happens.